Reduce inclusion of operations.hpp and depth_first_traversal.hpp
